HUM-145: code review fixes
