+++ packages/apps/human-app/frontend/src/modules/worker/hooks/use-jobs-notifications.tsx
@@ -9,7 +9,7 @@ export const useJobsNotifications = () => {
     showNotification({
       message: t('worker.jobs.successFullyAssignedJob'),
       type: 'success',
-      duration: 5000,
+      durationMs: 5000,
     });
   };
 
@@ -17,7 +17,7 @@ export const useJobsNotifications = () => {
     showNotification({
       message: getErrorMessageForError(error),
       type: 'warning',
-      duration: 5000,
+      durationMs: 5000,
     });
   };
