+++ packages/apps/human-app/frontend/src/shared/helpers/exhaustive-matching-guard.ts
@@ -1,0 +1,3 @@
+export const exhaustiveMatchingGuard = (_: never): never => {
+  throw new Error(`Should not have reached here`);
+};
