+++ packages/apps/human-app/frontend/src/shared/hooks/use-notification.tsx
@@ -1,13 +1,20 @@
 import { useCallback } from 'react';
 import { useSnackbar, type SnackbarKey } from 'notistack';
 import CloseIcon from '@mui/icons-material/Close';
-import { type ShowNotifProps } from '@/shared/types/notifications';
 import { useColorMode } from '@/shared/hooks/use-color-mode';
 import { colorPalette as lightColorPalette } from '@/shared/styles/color-palette';
 import { breakpoints } from '@/shared/styles/breakpoints';
 import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 
-const AUTO_HIDE_NOTIFICATION = 6000;
+export type TopNotificationType = 'success' | 'warning';
+
+export interface ShowNotifProps {
+  message: string;
+  type: TopNotificationType;
+  durationMs?: number;
+}
+
+const AUTO_HIDE_NOTIFICATION_MS = 116000;
 const FONT_COLOR = lightColorPalette.white;
 
 export const useNotification = () => {
@@ -20,6 +27,9 @@ export const useNotification = () => {
       <CloseIcon
         sx={{
           width: '20px',
+          position: 'absolute',
+          top: '15px',
+          right: isMobile ? '5px' : '38px',
           height: '20px',
           zIndex: 1,
           cursor: 'pointer',
@@ -30,17 +40,21 @@ export const useNotification = () => {
         }}
       />
     ),
-    [closeSnackbar]
+    [closeSnackbar, isMobile]
   );
 
   const showNotification = useCallback(
-    ({ message, type, duration = AUTO_HIDE_NOTIFICATION }: ShowNotifProps) => {
+    ({
+      message,
+      type,
+      durationMs = AUTO_HIDE_NOTIFICATION_MS,
+    }: ShowNotifProps) => {
       enqueueSnackbar(message, {
         variant: type,
-        autoHideDuration: duration,
+        autoHideDuration: durationMs,
         style: {
-          width: `calc(100svw - ${isMobile ? '32px' : '348px'})`,
-          marginRight: isMobile ? '18px' : '36px',
+          width: `calc(100svw - ${isMobile ? '32px' : '344px'})`,
+          marginRight: isMobile ? '18px' : '32px',
           backgroundColor:
             type === 'success'
               ? colorPalette.success.main
