+++ packages/apps/human-app/frontend/src/shared/providers/notifications-provider.tsx
@@ -1,6 +1,10 @@
 import { type ReactNode } from 'react';
 import { SnackbarProvider } from 'notistack';
-import { getNotificationIconByType } from '@/shared/helpers/get-notification-icon-by-type';
+import CheckCircleIcon from '@mui/icons-material/CheckCircle';
+import ErrorIcon from '@mui/icons-material/Error';
+import { type TopNotificationType } from '@/shared/hooks/use-notification';
+import { colorPalette as lightColorPalette } from '@/shared/styles/color-palette';
+import { exhaustiveMatchingGuard } from '@/shared/helpers/exhaustive-matching-guard';
 
 interface NotificationProviderProps {
   children: ReactNode;
@@ -9,6 +13,23 @@ interface NotificationProviderProps {
 
 const MAX_NOTIFICATIONS_VISIBLE = 5;
 
+const getNotificationIconByType = (
+  type: TopNotificationType,
+  sx?: Record<string, string>
+) => {
+  switch (type) {
+    case 'success':
+      return <CheckCircleIcon sx={{ fill: lightColorPalette.white, ...sx }} />;
+
+    case 'warning':
+      return <ErrorIcon sx={{ fill: lightColorPalette.white, ...sx }} />;
+
+    default: {
+      exhaustiveMatchingGuard(type);
+    }
+  }
+};
+
 export function NotificationProvider({
   children,
   maxSnacks = MAX_NOTIFICATIONS_VISIBLE,
@@ -16,7 +37,6 @@ export function NotificationProvider({
   return (
     <SnackbarProvider
       maxSnack={maxSnacks}
-      preventDuplicate
       anchorOrigin={{
         vertical: 'top',
         horizontal: 'right',
