+++ packages/apps/human-app/frontend/src/shared/styles/breakpoints.ts
@@ -1,4 +1,4 @@
 export const breakpoints = {
-  mobile: '@media (maxWidth:900px)',
+  mobile: '@media (maxidth:900px)',
   tablet: '@media (maxWidth:1200px)',
 };
