--- packages/apps/human-app/frontend/src/shared/types/notifications.ts
@@ -1,7 +1,0 @@
-export type TopNotificationType = 'success' | 'warning';
-
-export interface ShowNotifProps {
-  message: string;
-  type: TopNotificationType;
-  duration?: number;
-}
